added category sorting and course styling

+++ project.py
@@ -1,6 +1,5 @@
 import sys
 import os
-import urllib
 from flask import Flask, render_template, g, request, redirect, url_for, flash, jsonify, session
 from flask.ext.github import GitHub
 from werkzeug import secure_filename
@@ -45,6 +44,16 @@ def courses():
         this_user = None
     return render_template('courses.html', courses=all_courses, user=this_user, users=users)
 
+@app.route('/courses/<course_category>/')
+def category(course_category):
+    all_courses = db.query(Course).filter_by(category=course_category).all()
+    users = db.query(User).all()
+    if 'user_id' in session:
+        this_user = db.query(User).filter_by(id=session['user_id']).first()
+    else:
+        this_user = None
+    return render_template('category.html', courses=all_courses, user=this_user, users=users)
+
 @app.route('/new', methods=['GET', 'POST'])
 def newCourse():
     this_user = db.query(User).filter_by(id=session['user_id']).first()
@@ -60,7 +69,7 @@ def newCourse():
                                 category=request.form['category'], user_id=this_user.id)
         db.add(thisCourse)
         db.commit()
-        flash("New Course Created")
+        flash("New Course Created", "success")
         return redirect(url_for('courses'))
     else:
         return render_template('new_course.html')
@@ -73,34 +82,49 @@ def course(course_id):
             this_user = db.query(User).filter_by(id=session['user_id']).first()
         else:
             this_user = None
-        print(this_course.user_id)
-        print(this_user.id)
         return render_template('course.html', course=this_course, items=items, user=this_user)
 
 @app.route('/courses/<int:course_id>/edit', methods=['GET', 'POST'])
 def editCourse(course_id):
         this_course = db.query(Course).filter_by(id=course_id).one()
         old_pic = this_course.picture_name
-        if request.method == 'POST':
-            if request.form['name']:
-                this_course.name = request.form['name']
-            if request.form['description']:
-                this_course.description = request.form['description']
-            if request.form['category']:
-                this_course.category = request.form['category']
-            if request.files['file']:
-                if old_pic:
-                    os.remove(os.path.join(app.config['UPLOAD_FOLDER'], old_pic))
-                file = request.files['file']
-                filename = secure_filename(file.filename)
-                file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-                this_course.picture_name = filename
+        if checkAuth(course_id):
+            if request.method == 'POST':
+                if request.form['name']:
+                    this_course.name = request.form['name']
+                if request.form['description']:
+                    this_course.description = request.form['description']
+                if request.form['category']:
+                    this_course.category = request.form['category']
+                if request.files['file']:
+                    if old_pic:
+                        os.remove(os.path.join(app.config['UPLOAD_FOLDER'], old_pic))
+                    file = request.files['file']
+                    filename = secure_filename(file.filename)
+                    file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+                    this_course.picture_name = filename
 
-            db.add(this_course)
-            db.commit()
-            return redirect(url_for('course', course_id=course_id))
+                db.add(this_course)
+                db.commit()
+                flash("Course Edited", "success")
+                return redirect(url_for('courses'))
+            else:
+                return render_template('edit.html', course=this_course)
         else:
-            return render_template('edit.html', course=this_course)
+            flash("Access Denied", "danger")
+            return redirect(url_for('courses'))
+
+@app.route('/courses/<int:course_id>/deletecourse/', methods=['GET', 'POST'])
+def deleteCourse(course_id):
+        checkAuth(course_id)
+        this_course = db.query(Course).filter_by(id=course_id).one()
+        course_items = db.query(CourseItem).filter_by(course_id=course_id).all()
+        for i in course_items:
+            db.delete(i)
+        db.delete(this_course)
+        db.commit()
+        flash("Course Deleted", "danger")
+        return redirect(url_for('courses'))
 
 @app.route('/courses/<int:course_id>/newitem/', methods=['GET', 'POST'])
 def newItem(course_id):
@@ -122,43 +146,48 @@ def newItem(course_id):
                                   category=request.form['category'], text=request.form['text'], user_id=this_user.id)
         db.add(thisItem)
         db.commit()
-        flash("New Item Created")
+        flash("New Item Created", "success")
         return redirect(url_for('course', course_id=course_id))
     else:
         return render_template('new_item.html', course_id=course_id)
 
 @app.route('/courses/<int:course_id>/<int:item_id>/edititem/', methods=['GET', 'POST'])
 def editItem(course_id, item_id):
-        checkAuth(course_id)
         this_course = db.query(Course).filter_by(id=course_id).one()
         this_item = db.query(CourseItem).filter_by(id=item_id).one()
-        if request.method == 'POST':
-            if request.form['name']:
-                this_item.name = request.form['name']
-            if request.form['description']:
-                this_item.description = request.form['description']
-            if request.form['category']:
-                this_item.category = request.form['category']
-            if request.form['videourl']:
-                thisUrl = request.form['videourl']
-                referb = thisUrl.split("=")
-                this_item.youtube_url = referb[1]
-                this_item.audio_url = None
-                this_item.text = None
-            if request.form['audiourl']:
-                this_item.audio_url = request.form['audiourl']
-                this_item.youtube_url = None
-                this_item.text = None
-            if request.form['text']:
-                this_item.text = request.form['text']
-                this_item.audio_url = None
-                this_item.youtube_url = None
+        if checkAuth(course_id):
+            if request.method == 'POST':
+                if request.form['name']:
+                    this_item.name = request.form['name']
+                if request.form['description']:
+                    this_item.description = request.form['description']
+                if request.form['category']:
+                    this_item.category = request.form['category']
+                if request.form['videourl']:
+                    thisUrl = request.form['videourl']
+                    referb = thisUrl.split("=")
+                    this_item.youtube_url = referb[1]
+                    this_item.audio_url = None
+                    this_item.text = None
+                if request.form['audiourl']:
+                    this_item.audio_url = request.form['audiourl']
+                    this_item.youtube_url = None
+                    this_item.text = None
+                if request.form['text']:
+                    this_item.text = request.form['text']
+                    this_item.audio_url = None
+                    this_item.youtube_url = None
 
-            db.add(this_item)
-            db.commit()
-            return redirect(url_for('course', course_id=course_id))
+                db.add(this_item)
+                db.commit()
+                flash("Item Edited", "success")
+                return redirect(url_for('course', course_id=course_id))
+            else:
+                return render_template('edit_item.html', course=this_course, item=this_item)
         else:
-            return render_template('edit_item.html', course=this_course, item=this_item)
+            flash("Access Denied", "danger")
+            return redirect(url_for('courses'))
+
 
 @app.route('/courses/<int:course_id>/<int:item_id>/deleteitem/', methods=['GET', 'POST'])
 def deleteItem(course_id, item_id):
@@ -166,6 +195,7 @@ def deleteItem(course_id, item_id):
         this_item = db.query(CourseItem).filter_by(id=item_id).one()
         db.delete(this_item)
         db.commit()
+        flash("Item Deleted", "danger")
         return redirect(url_for('course', course_id=course_id))
 
 @app.route('/login')
@@ -200,18 +230,20 @@ def profile():
     else:
         return redirect(url_for('login'))
 
+@app.route('/logout')
+def logout():
+    session.pop('user_id', None)
+    return redirect(url_for('courses'))
 
 def checkAuth(course_id):
     if 'user_id' in session:
         this_course = db.query(Course).filter_by(id=course_id).one()
         if this_course.user_id == session['user_id']:
-            return
+            return True
         else:
-            flash("You don't have permission to edit that.")
-            return redirect(url_for('course', course_id=course_id))
+            return False
     else:
-        flash("Please Log in")
-        return redirect(url_for('courses'))
+        return False
 
 @github.access_token_getter
 def token_getter():
